Added option for output directory

+++ extract_qc.sh
@@ -4,11 +4,12 @@ set -eo pipefail
 trap 'exit_status="$?" && echo Failed on line: $LINENO at command: $BASH_COMMAND && echo "exit status $exit_status"' ERR
 
 #check that the script was called with 1 argument
-if [ $# -gt 2 ]
+if [ $# -gt 4 ]
 then
                 echo "
-                    Usage: bash $0 -d <experiment_directory> -j <json_filepath>
+                    Usage: bash $0 -d <experiment_directory> -j <json_filepath> <outputdir>
                     Either -d or -j are required but both can not be used.
+                    -o <output directory> is required
                     "
                 exit 1
 fi
@@ -18,7 +19,7 @@ experiment_dir=""
 json_file=""
 
 # Hantera flaggor med getopts
-while getopts "d:j:" opt; do
+while getopts "d:j:o:" opt; do
     case $opt in
         d)
             experiment_dir="$OPTARG"
@@ -33,8 +34,9 @@ while getopts "d:j:" opt; do
                 fi
             else
                 echo "
-                    Usage: bash $0 -d <experiment_directory> -j <json_filepath>
+                    Usage: bash $0 -d <experiment_directory> -j <json_filepath> <-o output directory>
                     Either -d or -j are required
+                    $experiment_dir was used with the -d flag but it is not a directory
                     "
                 exit 1
             fi
@@ -45,34 +47,50 @@ while getopts "d:j:" opt; do
             my_json="$json_file"
             if [ -f "$json_file" ]; then
                 my_json="$json_file"
-                real_experiment_path=$(realpath $(dirname "$my_json"))
+                real_experiment_path=$(realpath "$(dirname "$my_json")")
             else
                 echo "
-                    Usage: bash $0 -d <experiment_directory> -j <json_filepath>
+                    Usage: bash $0 -d <experiment_directory> -j <json_filepath> <-o output directory>
                     Either -d or -j are required
-                    $json_file is not a file
+                    -o is required.
+                    $json_file which you supplied with the -j flag is not a file.
                     "
                 exit 1
             fi
             ;;
+        o)
+            outputdir="$OPTARG"
+            echo "outputdir $outputdir"
+            ;;
         *)
             echo "
-                 Usage: bash $0 -d <experiment_directory> -j <json_filepath>
+                 Usage: bash $0 -d <experiment_directory> -j <json_filepath> <-o output directory>
                  Either -d or -j are required
+                 -o is required
                  "
             exit 1
             ;;
     esac
 done
 
-# Check that only one flag is used
-if [[ -z "$experiment_dir" && -z "$json_file" ]]; then
+# Check that either of experiment dir or json_file is defined by the user. Also check that the user specified an output directory
+if [[ (-z "$experiment_dir" && -z "$json_file") || -z "$outputdir" ]]; then
   echo "
         Error.
         Usage: bash $0 -d <experiment_directory> -j <json_filepath>
-        Either -d o -j are required
+        Either -d or -j are required
+        -o <output directory> is required
         "
   exit 1
+fi
+
+# create the results dir if the parent to the results dir exist and if needed.
+parent_outputdir=$(dirname "$outputdir")
+if [ -d "$parent_outputdir" ];then
+    mkdir -p "$outputdir"
+else
+    echo "The parent dir to $outputdir does not exist. The results directory could therefor not be created. Please correct this"
+    exit 1
 fi
 
 convert_qc_path=$(dirname $0)
@@ -109,9 +127,22 @@ cat "extract_qc_temp/extract_qc_temp_3.txt"
 echo ""
 run_name=$(grep "protocol_group_id" "extract_qc_temp/extract_qc_temp_3.txt" | cut -d "," -f 2)
 echo "Starting convert_qc.py"
-python "${convert_qc_path}"/convert_qc.py "extract_qc_temp/extract_qc_temp_3.txt" "convert_qc_output_${run_name}.tsv"
+python "${convert_qc_path}"/convert_qc.py "extract_qc_temp/extract_qc_temp_3.txt" "extract_qc_temp/convert_qc_output_${run_name}.tsv"
 echo "convert_qc.py finished"
 echo ""
+echo "moving results file"
+if [ ! -f "extract_qc_temp/convert_qc_output_${run_name}.tsv" ];then
+    echo "Results file could no be created. Exiting script"
+    exit 1
+fi
+mv "extract_qc_temp/convert_qc_output_${run_name}.tsv" "$outputdir"
+if [ -f "${outputdir}/convert_qc_output_${run_name}.tsv" ];then
+    echo "${outputdir}/convert_qc_output_${run_name}.tsv was created"
+else
+    echo "${outputdir}/convert_qc_output_${run_name}.tsv could not be created" 
+    exit 1
+fi
+
 echo "Removing temporary files"
 rm -r  extract_qc_temp
 echo "End of script"
